Worked on Larceny
Added the Move Selector even though it's not needed for this Quirk but I want consistency. I've also added conditions the steal move since the main weakness is "unable to use the Quirk on objects that are outside the line of sight, meaning that the user is left completely defenseless if their eyes become blinded."

This said, I have to find a better brightness value for this.